Actualización de las clases de repositories

+++ project/backend/src/main/java/com/jpa/backend/repositories/CentroRepository.java
@@ -2,11 +2,7 @@ package com.jpa.backend.repositories;
 
 import com.jpa.backend.entities.Centro;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-import java.util.List;
-
-public interface CentroRepository extends JpaRepository<Centro, Long> {
+public interface CentroRepository extends JpaRepository<Centro, Integer> {
 
 }
